requirements.txt
gradio
pandas
PyMuPDF
google-generativeai
